+++ src/services/database-integration.ts
@@ -3,6 +3,7 @@ import { SettingsService } from './settings-service';
 import { Conversation, Message, ConversationFolder, MessageContent, MessageContentType, FileJsonData } from '../types/chat';
 import { v4 as uuidv4 } from 'uuid';
 import { FileData } from '../types/file';
+import { ImageGenerationResult } from '../types/image';
 
 const SYSTEM_MESSAGE_CONTENT: MessageContent[] = [
     {
@@ -349,6 +350,17 @@ export class DatabaseIntegrationService {
             ...dbMessage,
         };
     }
+
+    public async saveImageGenerationResult(imageResult: ImageGenerationResult){
+        try {
+            const resultID = await this.dbService.saveImageGenerationResult(imageResult);
+            return resultID;
+        } catch (error) {
+            console.error('Error getting files:', error);
+            return null;
+        }
+    }
+
 
     public async saveFile(fileData: FileJsonData, arrayBuffer: ArrayBuffer): Promise<string> {
         const fileId = uuidv4();
