+++ src/types/chat.ts
@@ -53,16 +53,3 @@ export interface ConversationFolder{
   updatedAt: Date;
   colorFlag: string;
 }
-
-export interface ImageGenerationResult {
-  id: string;
-  prompt: string;
-  negativePrompt: string;
-  seed: string;
-  number: number;
-  status: string;
-  aspectRatio: string;
-  provider: string;
-  model: string;
-  images: MessageContent[];
-}
