+++ src/types/image.ts
@@ -1,0 +1,14 @@
+import { MessageContent } from "./chat";
+
+export interface ImageGenerationResult {
+    imageResultId: string;
+    prompt: string;
+    negativePrompt: string;
+    seed: string;
+    number: number;
+    status: string;
+    aspectRatio: string;
+    provider: string;
+    model: string;
+    images: MessageContent[];
+}
